refactored topnav select into one component
