--- client/src/components/LoggedInTopNavClick/LoggedInTopNavClick.css
@@ -1,7 +1,0 @@
-.logged-in-topnav-click{
-    grid-row: 1;
-    grid-column: 2;
-    background-color: blue;
-
-    display: flex;
-}
